+++ tests/Helper/UserWithAccessors.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Test\Helper;
+
+use Neat\Object\Accessors;
+
+class UserWithAccessors extends User
+{
+    use Accessors;
+}